requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Django==3.1.4
django-bootstrap-form==3.4
django-bootstrap4==2.3.1
idna==3.2
importlib-metadata==2.1.1
joblib==1.0.1
lxml==4.6.3
mysqlclient==2.0.3
nltk==3.6.3
pdfminer==20191125
Pillow==8.1.0
pycryptodome==3.10.4
python-docx==0.8.11
pytz==2020.4
regex==2021.9.30
requests==2.26.0
soupsieve==2.0.1
sqlparse==0.4.1
tqdm==4.62.3
urllib3==1.26.7
zipp==3.4.0